Use immutable (frozen) dataclasses and collection types
Immutable data avoids unintended non-local side effects. Dataclasses can
be "frozen" to achieve this, but lists and dictionaries are inherently
mutable. By using `Sequence` and `Mapping`, we can ensure that mutation
does not happen. This complicates the use of the dataclasses-json
package, unfortunately.

This commit also renames `Interval` to `Stint`.