Use timezone-aware datetime types
Before, the plain `datetime.now()` produced 'naive' datetime objects
which were implicitly assumed to be in the local (system) timezone
when serializing as JSON. The deserialized objects will always
be timezone-aware, also choosing the system timezone. This caused
serialized and deserialized objects to be different, even when
the system timezone stays the same. This is fixed for now by using
timezone-aware datetimes throughout.

There's a remaining issue, namely that a serialization roundtrip will
result in an aware datetime object but in a different timezone if the
original used a timezone other than the system's. This is not much of a
problem in practice, because aware type will always still reference the
same point in time, even if its representation changes. For unit tests,
this might be a problem, however, as those are using CEST iso timestamps
right now and will start to fail when run on a system with a different
timezone. Will fix later.