Use `tuple` instead of `Sequence` to record stints within an `Activity`
Tuples are truly immutable, so we might use those instead. This will
also remove some of the ambiguity regarding the concrete types when
imposing a fixed order in the next commit.

Tuples are also directly supported by dataclasses-json which allows us
to delete a bunch of code which was needed to add support for serializing
Sequence types.