Ensure the stints in an `Activity` are sorted
Currently, users can only produce unsorted activities by changing the
system clock. However, this will change when the `edit` subcommand is
implemented.

To ensure that activities remain sorted, a `__post_init__` method has
been added. Since the `stints` tuple is immutable, it is sufficient to
check (and enforce) that this invariant is upheld upon construction.
Since the dataclass is frozen, we cannot normally replace the member with
the sorted tuple, so we need to bypass the dataclass's `__setattr__`
implementation.