Implement the `edit` subcommand
This command can be used to make changes to an activity, e.g. to tweak
some of the times, add and remove stints, or even change the issue and
description.

When called on an activity, its `__str__` representation will be opened
in an editor. After changes have been made, the edited version will
be deserialized and, if successful, replaces the activity as a whole.

The editor is chosen according to click's heuristics but can be specified
in the config by adding the "editor" key.