Fix CI build by "exposing" `_data_dir()` as `data_dir()`
In CI, we use stubs for static analysis. Pyright does not stub functions
which are "private" (i.e. begin with a leading underscore) as those are
not part of the public API.

Usually, that is desirable as the stubs will not have any implementation
details, so those private functions won't be called within the stub
files. However, since `_data_dir()` is called from within a decorator
invocation and those are not stubbed out, we got a false positive.

Arguably, the decorators are in fact part of the module's public API,
because they potentially affect the functions' signatures. Therefore,
one could make the case that `_data_dir()` should indeed have been
public all along. Definitely not rationalizing my quick-and-dirty CI
fix here...